+++ src/scenes/elements.js
@@ -46,5 +46,26 @@ export class elements extends Phaser.Scene{
         scene.add.image(0, 0, bg).setOrigin(0,0).setDisplaySize(width,height);
     }
 
+    drawLoadingScreen(scene){
+        //  We loaded this image in our Boot Scene, so we can display it here
+        scene.add.image(0, 0, "bg_preload").setOrigin(0,0).setDisplaySize(width,height);
+        //  A simple progress bar. This is the outline of the bar.
+        scene.add.rectangle(mid_w, mid_h+(mid_h/2), 468, 32).setStrokeStyle(1, 0xffffff);
+        //  This is the progress bar itself. It will increase in size from the left based on the % of progress.
+        const bar = scene.add.rectangle(mid_w-230, mid_h+(mid_h/2), 4, 28, 0xffffff);
+        //  Use the 'progress' event emitted by the LoaderPlugin to update the loading bar
+        let displayedProgress = 0;
+
+        scene.load.on('progress', (progress) => {
+            // Actualiza la barra suavemente hacia el valor de "progress"
+            displayedProgress += (progress - displayedProgress) * 0.1; // Ajusta 0.1 para cambiar la velocidad
+
+            // Calcula el ancho de la barra basado en el progreso interpolado
+            bar.width = 4 + (460 * displayedProgress);
+
+            // Puedes usar una función como setTimeout o requestAnimationFrame si necesitas animarlo aún más
+        });
+    }
+
   
 }
